Pre-commit cleaned.
